changed cleanedNameMap and added nameToIDMap to replace getIDfromName method

+++ proj3/src/main/java/GraphDB.java
@@ -69,7 +69,7 @@ public class GraphDB {
             return this.lon;
         }
 
-        public double getID() {
+        public long getID() {
             return this.id;
         }
     }
